+++ lib/src/smart_arg.dart
@@ -456,12 +456,24 @@ class SmartArg {
     final b = a.type as ClassMirror;
     final command = b.newInstance('', []) as SmartArgCommand;
 
+    beforeCommandParse(command, arguments);
     command.parse(arguments);
+    beforeCommandExecute(command);
     command.execute(this);
+    afterCommandExecute(command);
   }
 
   void _resetParser() {
     _wasSet = {};
     _extras = [];
   }
+
+  /// Invoked before a command is parsed
+  void beforeCommandParse(SmartArgCommand command, List<String> arguments) {}
+
+  /// Invoked before a command is executed
+  void beforeCommandExecute(SmartArgCommand command) {}
+
+  /// Invoked after a command is executed
+  void afterCommandExecute(SmartArgCommand command) {}
 }