+++ test/smart_arg/command_parsing_test.dart
@@ -40,6 +40,26 @@ class TestSimpleCommand extends SmartArg {
 
   @Command(help: 'Get a file from a remote host')
   GetCommand? get;
+
+  List<String> hookOrder = [];
+
+  @override
+  void beforeCommandParse(SmartArgCommand command, List<String> arguments) {
+    super.beforeCommandParse(command, arguments);
+    hookOrder.add("beforeCommandParse");
+  }
+
+  @override
+  void beforeCommandExecute(SmartArgCommand command) {
+    super.beforeCommandExecute(command);
+    hookOrder.add("beforeCommandExecute");
+  }
+
+  @override
+  void afterCommandExecute(SmartArgCommand command) {
+    super.beforeCommandExecute(command);
+    hookOrder.add("afterCommandExecute");
+  }
 }
 
 void main() {
@@ -49,6 +69,18 @@ void main() {
     group('commands', () {
       setUp(() {
         whatExecuted = null;
+      });
+
+      test('calls hooks', () {
+        var cmd = TestSimpleCommand();
+        final args = cmd..parse(['get', '--filename=download.txt']);
+        expect(args.verbose, null);
+        expect(whatExecuted, 'get-command: download.txt');
+        expect(cmd.hookOrder, [
+          "beforeCommandParse",
+          "beforeCommandExecute",
+          "afterCommandExecute"
+        ]);
       });
 
       test('executes with no arguments', () {